🔨 fix: update app to use new AppConfig stuct
- remove redundant implementations in db module
- rename all the function calls
- TODO: fix tests/move unit tests to integration tests
