+++ src/bin/mathing.rs
@@ -2,8 +2,7 @@ use mathing_tui::prelude::*;
 
 #[tokio::main]
 async fn main() -> Result<()> {
-    Config::get_config();
-    StoreTotal::get_or_try_init().await?;
+    AppConfig::try_init().await?;
 
     let mut app = App::new()?;
     let app_result = app.run().await;
